refactor: rename mesh.kinds to mesh.connectivity, use **kwargs, update notebook
- Rename wmg.create.mesh.kinds -> wmg.create.mesh.connectivity
  (clearer separation between coordinate creation and connectivity)
- Replace explicit 'pattern' param with **kwargs in
  create_flat_singlescale_from_coordinates and
  create_flat_multiscale_from_coordinates (future-proofs for
  triangular and other mesh layouts)
- Update imports in base.py and tests/test_mesh_layout.py
- Update docs/creating_the_graph.ipynb to use new mesh_layout API
- Fix black and isort formatting
